fix bug with balance, fix tests

+++ TransactionStore.DataLayer/Repository/Interfaces/ITransactionRepository.cs
@@ -6,10 +6,10 @@ namespace TransactionStore.DataLayer.Repository
     {
         Task<long> AddTransaction(TransactionDto transaction);
         Task<List<long>> AddTransfer(TransferDto transaction);
-        Task<decimal> GetAccountBalance(long id);
+        Task<decimal> GetAccountBalance(int id);
         Task<TransactionDto> GetTransactionById(long id);
-        Task<List<TransactionDto>> GetTransactionsByAccountId(long id);
-        Task<List<TransactionDto>> GetTransactionsByAccountIdMinimal(long id);
-        Task<List<TransactionDto>> GetTransactionsByAccountIds(List<long> accountIds);
+        Task<List<TransactionDto>> GetTransactionsByAccountId(int id);
+        Task<List<TransactionDto>> GetTransactionsByAccountIdMinimal(int id);
+        Task<List<TransactionDto>> GetTransactionsByAccountIds(List<int> accountIds);
     }
 }
