+++ TransactionStore.DataLayer/Repository/TransactionRepository.cs
@@ -71,7 +71,7 @@ namespace TransactionStore.DataLayer.Repository
             return new List<long> { result.Item1, result.Item2 };
         }
 
-        public async Task<List<TransactionDto>> GetTransactionsByAccountId(long id)
+        public async Task<List<TransactionDto>> GetTransactionsByAccountId(int id)
         {
             _logger.LogInformation("Connecting to the database");
             using IDbConnection connection = Connection;
@@ -88,7 +88,7 @@ namespace TransactionStore.DataLayer.Repository
             return listTransactions;
         }
 
-        public async Task<List<TransactionDto>> GetTransactionsByAccountIdMinimal(long id)
+        public async Task<List<TransactionDto>> GetTransactionsByAccountIdMinimal(int id)
         {
             _logger.LogInformation("Connecting to the database");
             using IDbConnection connection = Connection;
@@ -105,14 +105,14 @@ namespace TransactionStore.DataLayer.Repository
             return listTransactions;
         }
 
-        public async Task<List<TransactionDto>> GetTransactionsByAccountIds(List<long> accountIds)
+        public async Task<List<TransactionDto>> GetTransactionsByAccountIds(List<int> accountIds)
         {
             _logger.LogInformation("Connecting to the database");
             using IDbConnection connection = Connection;
             _logger.LogInformation("Connection made");
 
             var tvpTable = new DataTable();
-            tvpTable.Columns.Add(new DataColumn("AccountId", typeof(long)));
+            tvpTable.Columns.Add(new DataColumn("AccountId", typeof(int)));
             accountIds.ForEach(id => tvpTable.Rows.Add(id));
 
             var listTransactions = (await connection.QueryAsync<TransactionDto>(
@@ -142,7 +142,7 @@ namespace TransactionStore.DataLayer.Repository
             return transactionDto;
         }
 
-        public async Task<decimal> GetAccountBalance(long id)
+        public async Task<decimal> GetAccountBalance(int id)
         {
             _logger.LogInformation("Connecting to the database");
             using IDbConnection connection = Connection;
